refactor(app): extract giant reducer dispatch matches into named handlers
Stream A of the readability branch — zero behavior change, adversarially
reviewed with normalized arm-vs-body diffs; clippy/tests green (3651 passed).

- reducer.rs App::dispatch: 17 handle_* methods; the 47-arm match is now
  one-line delegations (same shape as apply_media/apply_remote).
- keys.rs on_key: modal/global/mode routing helpers; reconstructed inline
  form is line-identical to the original.
- mouse.rs: on_mouse_click/on_mouse_target become thin dispatchers into new
  mouse_routing.rs (bodies moved verbatim); HitMap stays in mouse.rs. Four
  pub(super) helpers in mouse/{scrollbar,local_find}.rs widen to
  pub(in crate::app) for the sibling module.
- context_menu.rs execute_context_command: one helper per ContextTarget.
- recorder_reducer.rs on_recorder_event: dispatcher + 7 per-event helpers,
  statement order preserved.
- local_import.rs: pure row-detail/status formatting moves to local_format.rs
  as free functions over &ImportSessionRow.

Sizes: mouse.rs 1506->659 (+865 mouse_routing), local_import 1627->1357
(+301 to local_format), reducer 1063, keys 698.
